Fixed apply exception
Whoops! was returning lamType.t2 instead of argType.
So the point is, apply is superfluous for the linear part because it
says "just apply a function to an argument" but its important for the
untyped part because it says "do multiplication". the linear typechecker
already knows how to check application and the explicit apply is useless
info, so you just return the type of the first argument (which is the
function). the typechecker will then have App(argType, qubit) and it
knows how to check it thats correct!